+++ blocks/index.ts
@@ -1,2 +1,3 @@
 export * from "./generate-solo-nav-mesh";
 export * from "./generate-tiled-nav-mesh";
+export * from "./merge-positions-and-indices";